requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
pydantic
python-multipart
typing-extensions
python-dotenv